goextensions builder: export ModuleRoot
This follows the same pattern as LogToStdout and prevents a breaking
change to NewGoBuilder.

I debated creating a new Config type to be used as a parameter for
NewGoBuilder. But, in the end I wanted to avoid modifying the signature
of exposed methods. I am ok with this because the pattern for
LogToStdout is already used.